move wsdl parsing tests to their own file
